Added the Window module
This is just a wrapper around glfw for now. The odds of ever swapping glfw for our own WindowSystem are not that high for the foreseeable future.

N.B: The Window module is nowhere near complete; there are other functions that need to be created to satisfy the use cases for the window API. However, these functions will be added when needed so that no time is wasted while thinking about all the possible use cases for the window.

Signed-off-by: mo7sener <robear.selwans@outlook.com>