Stabilized Hierarchical Transform propagation
The unexpected behaviour mentioned in the previous commit was a product
of multiplying transformation matrices in a wrong order.

What made it hard to discover was the fact that the multiplication order
was wrong in two separate places, which caused a behaviour that was
close to the desired one but not quite the same.

Signed-off-by: Robear Selwans <robear.selwans@outlook.com>