Changed Event creation process
Due to gcc-10 forcing no-common, our current implementation of how
events are created for the engine was resulting in linker errors.

The process, unfortunately, was chanaged to a bit less user-friendly
one. However, This approach will probably be changed later when there is
time to think of a better solution.

Signed-off-by: Robear Selwans <robear.selwans@outlook.com>